Camp members will sit in user model. Prevent circular dependancy between camps model and users models
